Added EntryPoint code to ALUMApp.swift
